refactor chat formats
